requirements.txt
Pillow==11.0.0
tkinter
